fix: move declOfNum to utils, move carousel button to Pluralizer component
