+++ src/Complexes/Show/Carousel.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Grid, Row } from 'react-flexbox-grid';
 import styled from 'styled-components';
 import { imagesUrl } from './../../api';
+import Pluralizer from './Pluralizer';
 
 const Images = styled.div`
   display: flex;
@@ -14,27 +15,7 @@ const Image = styled.img`
   height: 400px;
 `;
 
-const Button = styled.button`
-  background-color: #00779a;
-  border: none;
-  color: #fff;
-  cursor: pointer;
-  display: block;
-  font-family: "Fira Sans", "Helvetica", sans-serif;
-  font-size: 0.625rem;
-  font-weight: 300;
-  line-height: 1.0;
-  margin-top: -2.625rem;
-  padding: 0.5rem 1rem;
-  position: absolute;
-`;
-
 const imagesSize = '512';
-
-function declOfNum(number, titles) {
-  const cases = [2, 0, 1, 1, 1, 2];
-  return `${number} ${titles[number % 100 > 4 && number % 100 < 20 ? 2 : cases[number % 10 < 5 ? number % 10 : 5]]}`;
-}
 
 export default props => (
   <div>
@@ -49,13 +30,7 @@ export default props => (
     </Images>
     <Grid>
       <Row>
-        <Button>
-          {declOfNum(props.images.length, [
-            'фотография',
-            'фотографии',
-            'фотографий',
-          ])}
-        </Button>
+        <Pluralizer imagesCount={props.images.length} />
       </Row>
     </Grid>
   </div>
